requirements.txt
pytest
requests
pytest-dependency
pytest-xdist
pytest-html
pylint
bandit
allure-pytest